# server/openapi/openapi.json
{
...
  "info": {
    "title": "Product Control Plane API",
    "version": "1.0.0-p1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/device-authorizations": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "browserFamily": {
                    "enum": [
                      "chrome",
                      "yandex_chromium"
                    ],
                    "type": "string"
                  },
                  "browserVersion": {
                    "maxLength": 64,
                    "minLength": 1,
                    "pattern": "^[A-Za-z0-9._+-]+$",
                    "type": "string"
                  },
                  "clientType": {
                    "enum": [
                      "browser_extension"
                    ],
                    "type": "string"
                  },
                  "deviceLabel": {
                    "maxLength": 128,
                    "minLength": 1,
                    "pattern": "^[^<>]*$",
                    "type": "string"
                  },
                  "extensionVersion": {
                    "maxLength": 64,
                    "minLength": 1,
                    "pattern": "^[A-Za-z0-9._+-]+$",
                    "type": "string"
                  }
                },
                "required": [
                  "clientType",
                  "browserFamily",
                  "extensionVersion"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "authorizationId": {
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                      "type": "string"
                    },
                    "deviceCode": {
                      "pattern": "^[A-Za-z0-9_-]{43}$",
                      "type": "string"
                    },
                    "expiresAt": {
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$",
                      "type": "string"
                    },
                    "status": {
                      "enum": [
                        "pending"
                      ],
                      "type": "string"
                    },
                    "userCode": {
                      "pattern": "^[ABCDEFGHJKLMNPQRSTUVWXYZ23456789]{4}-[ABCDEFGHJKLMNPQRSTUVWXYZ23456789]{4}$",
                      "type": "string"
                    }
                  },
                  "required": [
                    "status",
                    "authorizationId",
                    "deviceCode",
                    "userCode",
                    "expiresAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "enum": [
                            "INTERNAL_ERROR",
                            "INVALID_REQUEST",
                            "SERVICE_UNAVAILABLE",
                            "AUTH_RATE_LIMITED",
                            "AUTH_OTP_INVALID",
                            "AUTH_LOGIN_DENIED",
                            "AUTH_SESSION_INVALID",
                            "AUTH_CSRF_INVALID",
                            "DEVICE_AUTH_RATE_LIMITED",
                            "DEVICE_AUTH_INVALID",
                            "DEVICE_AUTH_FORBIDDEN",
                            "DEVICE_AUTH_STATE_CONFLICT",
                            "DEVICE_AUTH_IDEMPOTENCY_CONFLICT",
                            "DEVICE_AUTH_CLOSED"
                          ],
                          "type": "string"
                        },
                        "correlationId": {
                          "maxLength": 128,
                          "minLength": 1,
                          "pattern": "^[A-Za-z0-9._:-]+$",
                          "type": "string"
                        },
                        "message": {
                          "minLength": 1,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "enum": [
                            "INTERNAL_ERROR",
                            "INVALID_REQUEST",
                            "SERVICE_UNAVAILABLE",
                            "AUTH_RATE_LIMITED",
                            "AUTH_OTP_INVALID",
                            "AUTH_LOGIN_DENIED",
                            "AUTH_SESSION_INVALID",
                            "AUTH_CSRF_INVALID",
                            "DEVICE_AUTH_RATE_LIMITED",
                            "DEVICE_AUTH_INVALID",
                            "DEVICE_AUTH_FORBIDDEN",
                            "DEVICE_AUTH_STATE_CONFLICT",
                            "DEVICE_AUTH_IDEMPOTENCY_CONFLICT",
                            "DEVICE_AUTH_CLOSED"
                          ],
                          "type": "string"
                        },
                        "correlationId": {
                          "maxLength": 128,
                          "minLength": 1,
                          "pattern": "^[A-Za-z0-9._:-]+$",
                          "type": "string"
                        },
                        "message": {
                          "minLength": 1,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "enum": [
                            "INTERNAL_ERROR",
                            "INVALID_REQUEST",
                            "SERVICE_UNAVAILABLE",
                            "AUTH_RATE_LIMITED",
                            "AUTH_OTP_INVALID",
                            "AUTH_LOGIN_DENIED",
                            "AUTH_SESSION_INVALID",
                            "AUTH_CSRF_INVALID",
                            "DEVICE_AUTH_RATE_LIMITED",
                            "DEVICE_AUTH_INVALID",
                            "DEVICE_AUTH_FORBIDDEN",
                            "DEVICE_AUTH_STATE_CONFLICT",
                            "DEVICE_AUTH_IDEMPOTENCY_CONFLICT",
                            "DEVICE_AUTH_CLOSED"
                          ],
                          "type": "string"
                        },
                        "correlationId": {
                          "maxLength": 128,
                          "minLength": 1,
                          "pattern": "^[A-Za-z0-9._:-]+$",
                          "type": "string"
                        },
                        "message": {
                          "minLength": 1,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "enum": [
                            "INTERNAL_ERROR",
                            "INVALID_REQUEST",
                            "SERVICE_UNAVAILABLE",
                            "AUTH_RATE_LIMITED",
                            "AUTH_OTP_INVALID",
                            "AUTH_LOGIN_DENIED",
                            "AUTH_SESSION_INVALID",
                            "AUTH_CSRF_INVALID",
                            "DEVICE_AUTH_RATE_LIMITED",
                            "DEVICE_AUTH_INVALID",
                            "DEVICE_AUTH_FORBIDDEN",
                            "DEVICE_AUTH_STATE_CONFLICT",
                            "DEVICE_AUTH_IDEMPOTENCY_CONFLICT",
                            "DEVICE_AUTH_CLOSED"
                          ],
                          "type": "string"
                        },
                        "correlationId": {
                          "maxLength": 128,
                          "minLength": 1,
                          "pattern": "^[A-Za-z0-9._:-]+$",
                          "type": "string"
                        },
                        "message": {
                          "minLength": 1,
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message",
                        "correlationId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": []
}